[WIP] Reissue certs as part of the join of a cluster
Certs must be reissued as part of joining the cluster otherwise the k8s
workers will have problems authenticating and verifying the certificates
from the primary node. The reissue of certs can be disabled by using a
touch file (available in order to prevent a node from restarting
networking).

This fixes the joining a node to a cluster by temporarily removing the
touch file during the node join process and restores the touch file (if
necessary) after the join is done.

Signed-off-by: Billy Olsen <billy.olsen@gmail.com>